create network policy for fluentbit
