+++ internal/fluentbit/ports/ports.go
@@ -1,0 +1,7 @@
+package ports
+
+const (
+	HTTP       = 2020
+	Metrics    = 2021
+	IstioEnvoy = 15090
+)
