+++ internal/resources/common/resources.go
@@ -2,9 +2,11 @@ package common
 
 import (
 	corev1 "k8s.io/api/core/v1"
+	networkingv1 "k8s.io/api/networking/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
 func MakeServiceAccount(name types.NamespacedName) *corev1.ServiceAccount {
@@ -31,4 +33,58 @@ func MakeClusterRoleBinding(name types.NamespacedName) *rbacv1.ClusterRoleBindin
 		},
 	}
 	return &clusterRoleBinding
+}
+
+func MakeNetworkPolicy(name types.NamespacedName, allowedPorts []int32, labels map[string]string) *networkingv1.NetworkPolicy {
+	return &networkingv1.NetworkPolicy{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name.Name + "-pprof-deny-ingress",
+			Namespace: name.Namespace,
+			Labels:    labels,
+		},
+		Spec: networkingv1.NetworkPolicySpec{
+			PodSelector: metav1.LabelSelector{
+				MatchLabels: labels,
+			},
+			PolicyTypes: []networkingv1.PolicyType{
+				networkingv1.PolicyTypeIngress,
+				networkingv1.PolicyTypeEgress,
+			},
+			Ingress: []networkingv1.NetworkPolicyIngressRule{
+				{
+					From: []networkingv1.NetworkPolicyPeer{
+						{
+							IPBlock: &networkingv1.IPBlock{CIDR: "0.0.0.0/0"},
+						},
+					},
+					Ports: makeNetworkPolicyPorts(allowedPorts),
+				},
+			},
+			Egress: []networkingv1.NetworkPolicyEgressRule{
+				{
+					To: []networkingv1.NetworkPolicyPeer{
+						{
+							IPBlock: &networkingv1.IPBlock{CIDR: "0.0.0.0/0"},
+						},
+					},
+				},
+			},
+		},
+	}
+}
+
+func makeNetworkPolicyPorts(ports []int32) []networkingv1.NetworkPolicyPort {
+	var networkPolicyPorts []networkingv1.NetworkPolicyPort
+
+	tcpProtocol := corev1.ProtocolTCP
+
+	for idx := range ports {
+		port := intstr.FromInt32(ports[idx])
+		networkPolicyPorts = append(networkPolicyPorts, networkingv1.NetworkPolicyPort{
+			Protocol: &tcpProtocol,
+			Port:     &port,
+		})
+	}
+
+	return networkPolicyPorts
 }
